# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ingestion/trigger": {
      "post": {
        "operationId": "IngestionController_trigger",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerIngestionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Ingestion"
        ]
      }
    },
...
  },
  "info": {
    "title": "Document Management API",
    "description": "API for managing users, documents, and ingestion",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TriggerIngestionDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}